requirements.txt
numpy==1.25.2
pydantic==2.3.0
pydantic_core==2.6.3
pygame-ce==2.3.1
pillow==10.0.0
